integrations-examples/livekit-agent/pyproject.toml
[project]
name = "livekit-example-transcriber"
version = "0.1.0"
description = "Simple Gladia transcriber example using LiveKit Cloud and LiveKit Agents Playground"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "livekit-agents[gladia]>=1.2.6",
    "python-dotenv>=1.1.1",
]
